@@ -12,8 +12,9 @@ QEFIEntryDPDetailView::QEFIEntryDPDetailView(QEFIDevicePath *dp, QWidget *parent
         convert_device_path_subtype_to_name(dp->type(), dp->subType()))
     );
     m_topLevelLayout->addRow(QStringLiteral(""),
-        new QLabel(QString::asprintf("%02X %02X",
-            dp->type(), dp->subType()))
+        new QLabel(QStringLiteral("%1 %1")
+	    .arg(dp->type(), 2, 16, QLatin1Char('0'))
+	    .arg(dp->subType(), 2, 16, QLatin1Char('0')))
     );
     // Parse device path and add more properties
     QList<QPair<QString, QString>> attrs = convert_device_path_attrs(dp);
